+++ test-fail/test-fail.ts
@@ -1,0 +1,5 @@
+import { test } from 'purple-tape'
+
+test('shall fail', (t) => {
+    t.equal(1, 2, '1 shall be 2')
+})