+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Features
 
 - Add -e option for errors-only mode that only prints output from failing tests plus the summary
+- Show help message when run without arguments
 
 ## [1.8.0] 2025-11-14